+++ backend/demo/demo/src/main/java/testing/demo/Repository/AdminRepo.java
@@ -1,0 +1,10 @@
+package testing.demo.Repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import testing.demo.model.Admin;
+
+import java.util.Optional;
+
+public interface AdminRepo extends JpaRepository<Admin,Integer> {
+    Optional<Object> findById(Admin admin);
+}
