+++ backend/demo/demo/src/main/java/testing/demo/Repository/SchoolRepo.java
@@ -1,0 +1,7 @@
+package testing.demo.Repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import testing.demo.model.School;
+
+public interface SchoolRepo extends JpaRepository<School,Integer> {
+}
