--- demo/demo/src/main/java/testing/demo/Repository/UserCredentialRepo.java
@@ -1,9 +1,0 @@
-package testing.demo.Repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import testing.demo.model.UserCredentials;
-
-public interface UserCredentialRepo extends JpaRepository<UserCredentials, Integer> {
-
-
-}
